+++ app/components/features/quickstart/board-favorites/item/index.ts
@@ -24,7 +24,7 @@ export default class QuickstartBoardFavoriteComponent extends Component<Signatur
 
   @action handleLinkClick() {
     if (this.args.inSidebar && !this.renderer.isDesktop) {
-      this.renderer.closeLeftSidebar();
+      this.renderer.toggleLeftSidebar(false);
     }
   }