+++ app/components/features/quickstart/newsfeed/bookmark/index.ts
@@ -25,7 +25,7 @@ export default class QuickstartNewsfeedBookmarkComponent extends Component<Signa
 
   @action handleLinkClick() {
     if (this.args.inSidebar && !this.renderer.isDesktop) {
-      this.renderer.closeLeftSidebar();
+      this.renderer.toggleLeftSidebar(false);
     }
   }
 }